+++ packages/blast-core/cmd/blast-cmd.js
@@ -5,17 +5,16 @@ const { hideBin } = require('yargs/helpers')
 const commands = require('./commands')
 const BlastError = require('../utilities/blast-error')
 
-const { getTasks } = require('../lib/task')
+const { customTasks } = require('../utilities/task')
 const { getConfig } = require('../utilities/config-utils')
 
 
 async function main() {
+  
   if (hideBin(process.argv)[0] != 'init') {
     getConfig()
   }
-  const customTasks = getTasks()
-  // console.log(customTasks);
-  
+
   await yargs(hideBin(process.argv))
     .scriptName('blast')
     .usage('Usage: $0 <command> [arguments] [command options]')
